CRS-325 test removed to diagnose circle job failure
